Split solution into separate .NET project types. - Used file *linking* to keep the code in one location and the other - .NET 3.5 (Main code repo) - .NET 4.0 (code linked) - .NET Pcl (code linked)
