--- CreditCardUnitTest/Data/ValidCards.json
@@ -1,95 +1,0 @@
-﻿{
-    "Visa": [
-        "4012 8888 8888 1881",
-        "4111 1111 1111 1111",
-        "4222 2222 2222 2",
-        "4917 6100 0000 0000"
-    ],
-    "MasterCard": [
-        "5274 5763 9425 9961",
-        "5555 5555 5555 4444",
-        "5105 1051 0510 5100"
-    ],
-    "DinersClub": [
-        "3020 4169 3226 43",
-        "3021 8047 1965 57",
-        "3022 1511 5632 52",
-        "3600 0000 0000 08",
-        "3614 8900 6479 13",
-        "3670 0102 0000 00",
-        "3852 0000 0232 37",
-        "3056 9309 0259 04",
-        "3020 4169 3226 43"
-    ],
-    "AmericanExpress": [
-        "3714 4963 5398 431",
-        "3787 3449 3671 000",
-        "3400 0000 0000 009",
-        "3411 1111 1111 111",
-        "3434 3434 3434 343",
-        "3468 2763 0435 344",
-        "3700 0000 0000 002",
-        "3700 0020 0000 000",
-        "3704 0726 9909 809",
-        "3705 5601 9309 221",
-        "3714 4963 5398 431",
-        "3742 0000 0000 004",
-        "3764 6228 0921 451",
-        "3777 5274 9896 404",
-        "3782 8224 6310 005"
-    ],
-    "Discover": [
-        "6011 1111 1111 1117",
-        "6011 0009 9013 9424",
-        "6011 0000 0000 0004",
-        "6011 0004 0000 0000",
-        "6011 1532 1637 1980",
-        "6011 6011 6011 6611",
-        "6011 6874 8256 4166",
-        "6011 8148 3690 5651"
-    ],
-    "Maestro": [
-        "6759 6498 2643 8453",
-        "5641 8200 0000 0005",
-        "5033 9619 8909 17",
-        "5868 2416 0825 5333 38",
-        "6799 9901 0000 0000 019"
-    ],
-    "JCB": [
-        "3575 7591 5225 4876",
-        "3566 0020 2036 0505",
-        "1800 0163 8277 392",
-        "3569 9900 0000 0009",
-        "3530 1113 3330 0000",
-        "3572 6600 0000 0000 006"
-    ],
-    "Solo": [
-        "6767 6222 2222 2222 222",
-        "6334 5805 0000 0000",
-        "6334 9000 0000 0005",
-        "6334 7306 0000 0000 00",
-        "6767 6767 6767 6767 671"
-    ],
-    "ChinaUnionPay": [
-        "6264 1852 1292 2132 067",
-        "6288 9977 1545 2584",
-        "6269 9920 5813 4322"
-    ],
-    "Dankort": [
-        "5019 7170 1010 3742"
-    ],
-    "Switch": [
-        "6331 1019 9999 0016"
-    ],
-    "Laser": [
-        "6304 9506 0000 0000 00",
-        "6304 9000 1774 0292 441"
-    ],
-    "RuPay": [
-        "6076 6000 0619 9992",
-        "6070 5500 5000 0047"
-    ],
-    "Hipercard": [
-        "3841 0058 9908 8180 330"
-    ]
-}
