--- CreditCardUnitTest/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Newtonsoft.Json" version="7.0.1" targetFramework="net451" />
-</packages>
