+++ CreditCardValidator.sln
@@ -3,9 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
 VisualStudioVersion = 12.0.31101.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CreditCardValidator", "CreditCardValidator\CreditCardValidator.csproj", "{221006C3-0DE0-40FC-B395-1C59A2B468E4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CreditCardValidator.Net40", "Src\CreditCardValidator.Net40\CreditCardValidator.Net40.csproj", "{007F80D3-2EAF-483D-A9B5-39E862DFC25C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CreditCardUnitTest", "CreditCardUnitTest\CreditCardUnitTest.csproj", "{1D3840DE-41C6-4517-BE06-309AA4A73759}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CreditCardValidator.Net35", "Src\CreditCardValidator.Net35\CreditCardValidator.Net35.csproj", "{F138E4A6-CDC7-405C-B9FB-879ECD343E53}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CreditCardValidator.Pcl", "Src\CreditCardValidator.Pcl\CreditCardValidator.Pcl.csproj", "{10FCFCC0-BF9D-40AE-9482-37CB409A46AB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CreditCardValidator.Tests", "Src\CreditCardValidator.Tests\CreditCardValidator.Tests.csproj", "{1D3840DE-41C6-4517-BE06-309AA4A73759}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,10 +17,18 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{221006C3-0DE0-40FC-B395-1C59A2B468E4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{221006C3-0DE0-40FC-B395-1C59A2B468E4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{221006C3-0DE0-40FC-B395-1C59A2B468E4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{221006C3-0DE0-40FC-B395-1C59A2B468E4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{007F80D3-2EAF-483D-A9B5-39E862DFC25C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{007F80D3-2EAF-483D-A9B5-39E862DFC25C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{007F80D3-2EAF-483D-A9B5-39E862DFC25C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{007F80D3-2EAF-483D-A9B5-39E862DFC25C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F138E4A6-CDC7-405C-B9FB-879ECD343E53}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F138E4A6-CDC7-405C-B9FB-879ECD343E53}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F138E4A6-CDC7-405C-B9FB-879ECD343E53}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F138E4A6-CDC7-405C-B9FB-879ECD343E53}.Release|Any CPU.Build.0 = Release|Any CPU
+		{10FCFCC0-BF9D-40AE-9482-37CB409A46AB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{10FCFCC0-BF9D-40AE-9482-37CB409A46AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{10FCFCC0-BF9D-40AE-9482-37CB409A46AB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{10FCFCC0-BF9D-40AE-9482-37CB409A46AB}.Release|Any CPU.Build.0 = Release|Any CPU
 		{1D3840DE-41C6-4517-BE06-309AA4A73759}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1D3840DE-41C6-4517-BE06-309AA4A73759}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1D3840DE-41C6-4517-BE06-309AA4A73759}.Release|Any CPU.ActiveCfg = Release|Any CPU
