--- CreditCardValidator/Properties/AssemblyInfo.cs
@@ -1,34 +1,0 @@
-﻿using System.Reflection;
-
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("CreditCardValidator")]
-[assembly: AssemblyDescription("Credit Card Validator")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("CreditCardValidator")]
-[assembly: AssemblyCopyright("Copyright © 2015 Gustavo Frizzo")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-// Setting ComVisible to false makes the types in this assembly not visible 
-// to COM components.  If you need to access a type in this assembly from 
-// COM, set the ComVisible attribute to true on that type.
-//[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-//[assembly: Guid("b6b26040-dac5-42fa-8a96-da242582de48")]
-
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version 
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Build and Revision Numbers 
-// by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.3.0.*")]
-//[assembly: AssemblyFileVersion("1.2.1.*")]
