+++ Src/CreditCardValidator.Net35/Helpers/StringHelper.cs
@@ -1,0 +1,14 @@
+﻿using System.Linq;
+
+namespace CreditCardValidator.Helpers
+{
+    internal static class StringHelper
+    {
+        public static string RemoveWhiteSpace(this string input)
+        {
+            return new string(input.ToCharArray()
+                .Where(c => !char.IsWhiteSpace(c))
+                .ToArray());
+        }
+    }
+}
