+++ Src/CreditCardValidator.Net35/Helpers/ValidationHelper.cs
@@ -1,0 +1,17 @@
+﻿using System.Linq;
+
+namespace CreditCardValidator.Helpers
+{
+    internal static class ValidationHelper
+    {
+        public static bool IsAValidNumber(string number)
+        {
+            number = number.RemoveWhiteSpace();
+
+            return (number
+                .ToCharArray()
+                .All(char.IsNumber) &&
+                    !string.IsNullOrEmpty(number));
+        }
+    }
+}
