+++ Src/CreditCardValidator.Net40/CreditCardValidator.Net40.csproj
@@ -1,0 +1,84 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
+  <PropertyGroup>
+    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
+    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
+    <ProjectGuid>{007F80D3-2EAF-483D-A9B5-39E862DFC25C}</ProjectGuid>
+    <OutputType>Library</OutputType>
+    <AppDesignerFolder>Properties</AppDesignerFolder>
+    <RootNamespace>CreditCardValidator</RootNamespace>
+    <AssemblyName>CreditCardValidator</AssemblyName>
+    <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
+    <FileAlignment>512</FileAlignment>
+    <TargetFrameworkProfile />
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
+    <DebugSymbols>true</DebugSymbols>
+    <DebugType>full</DebugType>
+    <Optimize>false</Optimize>
+    <OutputPath>bin\Debug\</OutputPath>
+    <DefineConstants>DEBUG;TRACE</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
+    <DebugType>pdbonly</DebugType>
+    <Optimize>true</Optimize>
+    <OutputPath>bin\Release\</OutputPath>
+    <DefineConstants>TRACE</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <ItemGroup>
+    <Reference Include="System" />
+    <Reference Include="System.Core" />
+    <Reference Include="Microsoft.CSharp" />
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="..\CreditCardValidator.Net35\BrandInfo.cs">
+      <Link>BrandInfo.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\CreditCardData.cs">
+      <Link>CreditCardData.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\CreditCardDetector.cs">
+      <Link>CreditCardDetector.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\CreditCardFactory.cs">
+      <Link>CreditCardFactory.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\CreditCardStringExtension.cs">
+      <Link>CreditCardStringExtension.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\Helpers\StringHelper.cs">
+      <Link>Helpers\StringHelper.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\Helpers\ValidationHelper.cs">
+      <Link>Helpers\ValidationHelper.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\Luhn.cs">
+      <Link>Luhn.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\MajorIndustryIdentifier.cs">
+      <Link>MajorIndustryIdentifier.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\Rule.cs">
+      <Link>Rule.cs</Link>
+    </Compile>
+    <Compile Include="Properties\AssemblyInfo.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="..\CreditCardValidator.Net35\Data\Brands.json">
+      <Link>Data\Brands.json</Link>
+    </None>
+  </ItemGroup>
+  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
+  <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
+       Other similar extension points exist, see Microsoft.Common.targets.
+  <Target Name="BeforeBuild">
+  </Target>
+  <Target Name="AfterBuild">
+  </Target>
+  -->
+</Project>
