+++ Src/CreditCardValidator.Pcl/CreditCardValidator.Pcl.csproj
@@ -1,0 +1,83 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
+  <PropertyGroup>
+    <MinimumVisualStudioVersion>11.0</MinimumVisualStudioVersion>
+    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
+    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
+    <ProjectGuid>{10FCFCC0-BF9D-40AE-9482-37CB409A46AB}</ProjectGuid>
+    <OutputType>Library</OutputType>
+    <AppDesignerFolder>Properties</AppDesignerFolder>
+    <RootNamespace>CreditCardValidator</RootNamespace>
+    <AssemblyName>CreditCardValidator</AssemblyName>
+    <DefaultLanguage>en-US</DefaultLanguage>
+    <FileAlignment>512</FileAlignment>
+    <ProjectTypeGuids>{786C830F-07A1-408B-BD7F-6EE04809D6DB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
+    <TargetFrameworkProfile>Profile111</TargetFrameworkProfile>
+    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
+    <DebugSymbols>true</DebugSymbols>
+    <DebugType>full</DebugType>
+    <Optimize>false</Optimize>
+    <OutputPath>bin\Debug\</OutputPath>
+    <DefineConstants>DEBUG;TRACE</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
+    <DebugType>pdbonly</DebugType>
+    <Optimize>true</Optimize>
+    <OutputPath>bin\Release\</OutputPath>
+    <DefineConstants>TRACE</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <ItemGroup>
+    <!-- A reference to the entire .NET Framework is automatically included -->
+    <None Include="..\CreditCardValidator.Net35\Data\Brands.json">
+      <Link>Data\Brands.json</Link>
+    </None>
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="..\CreditCardValidator.Net35\BrandInfo.cs">
+      <Link>BrandInfo.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\CreditCardData.cs">
+      <Link>CreditCardData.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\CreditCardDetector.cs">
+      <Link>CreditCardDetector.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\CreditCardFactory.cs">
+      <Link>CreditCardFactory.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\CreditCardStringExtension.cs">
+      <Link>CreditCardStringExtension.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\Helpers\StringHelper.cs">
+      <Link>Helpers\StringHelper.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\Helpers\ValidationHelper.cs">
+      <Link>Helpers\ValidationHelper.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\Luhn.cs">
+      <Link>Luhn.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\MajorIndustryIdentifier.cs">
+      <Link>MajorIndustryIdentifier.cs</Link>
+    </Compile>
+    <Compile Include="..\CreditCardValidator.Net35\Rule.cs">
+      <Link>Rule.cs</Link>
+    </Compile>
+    <Compile Include="Properties\AssemblyInfo.cs" />
+  </ItemGroup>
+  <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
+  <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
+       Other similar extension points exist, see Microsoft.Common.targets.
+  <Target Name="BeforeBuild">
+  </Target>
+  <Target Name="AfterBuild">
+  </Target>
+  -->
+</Project>
